Moved Animation Controls methods in a dedicated Interface Component: AnimationControls
